+++ variables_load_balancer.tf
@@ -1,6 +1,6 @@
 ## Load balancer variables
 # Load balancer internal or external
-variable "lb_internal" {
+variable "create_internal_lb" {
   description = "Define if the loadbalancer should be internal or not"
   default     = true
 }
@@ -11,58 +11,18 @@ variable "lb_name" {
   default     = "loadbalancer"
 }
 
-# Load balancer public subnet groups
-variable "lb_public_subnet_ids" {
-  type        = "list"
-  description = "The Public Subnet ID(s) which should be attached to the loadbalancer"
-  default     = []
-}
-
-# Load balancer private subnet groups
-variable "lb_private_subnet_ids" {
-  type        = "list"
-  description = "The Private Subnet ID(s) which should be attached to the loadbalancer"
+# Load balancer subnet groups
+variable "lb_subnet_ids" {
+  type        = list(string)
+  description = "The subnet ID(s) which should be attached to the loadbalancer"
   default     = []
 }
 
 ## Load balancer security groups
-# Load balancer additional security groups
 variable "lb_security_group_ids" {
-  type        = "list"
+  type        = list(string)
   description = "Additional Security Group ID(s) which should be attached to the loadbalancer"
   default     = []
-}
-
-# Loadbalancer source traffic name (for security group allowing listener traffic in)
-variable "lb_source_traffic_name" {
-  description = "Name of source traffic for the loadbalancer within the HTTP and HTTPS security group. E.g. Rule-allow-${var.lb_source_traffic_name}-in-HTTP"
-  default     = ""
-}
-
-# Loadbalancer HTTP security group name
-variable "lb_sg_http_name" {
-  description = "Name of the HTTP security group"
-  default     = ""
-}
-
-# Loadbalancer HTTPS security group name
-variable "lb_sg_https_name" {
-  description = "Name of the HTTP security group"
-  default     = ""
-}
-
-# CIDR blocks specifying IP(s) allowed in on the http listener port for the loadbalancer
-variable "rule_allow_lb_http_listener_traffic_in_cidr_blocks" {
-  type        = "list"
-  description = "CIDR blocks loadbalancer http listener source traffic"
-  default     = ["0.0.0.0/0"]
-}
-
-# CIDR blocks specifying IP(s) allowed in on the https listener port for the loadbalancer
-variable "rule_allow_lb_https_listener_traffic_in_cidr_blocks" {
-  type        = "list"
-  description = "CIDR blocks loadbalancer https listener source traffic"
-  default     = ["0.0.0.0/0"]
 }
 
 # Load balancer idle timeout
@@ -77,59 +37,31 @@ variable "lb_enable_http2" {
   default     = true
 }
 
+# Load balancer logs bucket
+variable "lb_logs_bucket_enabled" {
+  description = "Boolean to enable / disable access_logs"
+  default     = true
+}
+
+variable "lb_logs_bucket_name" {
+  description = "The S3 bucket name to store the logs in"
+  default     = ""
+}
+
+variable "lb_logs_bucket_prefix" {
+  description = "The S3 bucket prefix. Logs are stored in the root if not configured"
+  default     = ""
+}
+
 # Load balancer deletion protection
 variable "lb_enable_deletion_protection" {
   description = "Enable deletion protection"
   default     = false
 }
 
-## Load balancer listener variables
-# Load balancer HTTP listener needed?
-variable "lb_http_listener" {
-  description = "If true add a HTTP listener"
-  default     = true
-}
-
-# Load balancer HTTPS listener needed?
-variable "lb_https_listener" {
-  description = "If true add a HTTPS listener"
-  default     = false
-}
-
-# Load balancer HTTPS offloading?
-variable "lb_https_offloading" {
-  description = "If true offload to HTTP"
-  default     = false
-}
-
-# Load balancer HTTP listener port
-variable "lb_http_listener_port" {
-  description = "HTTP listener port of the loadbalancer"
-  default     = 80
-}
-
-# Loadbalancer HTTPS listener port
-variable "lb_https_listener_port" {
-  description = "HTTPS listener port of the loadbalancer"
-  default     = 443
-}
-
-# HTTP host headers for load balancer listener rules
-variable "http_host_headers" {
-  type        = "list"
-  description = "Http ost headers for load balancer listener rules"
-  default     = []
-}
-
-# HTTPS host headers for load balancer listener rules
-variable "https_host_headers" {
-  type        = "list"
-  description = "Https ost headers for load balancer listener rules"
-  default     = []
-}
-
-# Certificate ARN for the HTTPS listener
-variable "certificate_arn" {
-  description = "Certificate ARN for the HTTPS listener"
-  default     = ""
+# Tags
+variable "lb_tags" {
+  description = "Tags for the load balancer"
+  type        = map(string)
+  default     = {}
 }