+++ variables_shared_target_group.tf
@@ -3,3 +3,9 @@ variable "target_ids" {
   description = "Instance IDs for the target group(s)"
   default     = ""
 }
+
+variable "tg_tags" {
+  description = "Tags for target groups"
+  type        = map(string)
+  default     = {}
+}